+++ benches/benches/bevy_ecs/observers/mod.rs
@@ -1,6 +1,8 @@
 use criterion::criterion_group;
 
 mod propagation;
+mod simple;
 use propagation::*;
+use simple::*;
 
-criterion_group!(observer_benches, event_propagation);
+criterion_group!(observer_benches, event_propagation, observe_simple);
